QmlProfiler: Fix compile warning on Windows.
c:\dev\insrc\devqc\src\plugins\qmlprofiler\qmlevent.h(274): warning
C4267: '=': conversion from 'size_t' to 'quint16', possible loss of data

+++ src/plugins/qmlprofiler/qmlevent.h
@@ -271,7 +271,7 @@ private:
     {
         Number *data;
         m_dataLength = squeezable<size_t, quint16>(numbers.size()) ?
-                    numbers.size() : std::numeric_limits<quint16>::max();
+                    static_cast<quint16>(numbers.size()) : std::numeric_limits<quint16>::max();
         if (m_dataLength > sizeof(m_data) / sizeof(Number)) {
             if (squeeze<Container, Number>(numbers))
                 return;
